:construction_worker: ci: Run test-tvl from CI
Detect the project of the pull request and run the `test-tvl` command
if relevant.

I've edited two adapters to demonstrate the feature.
The `test-tvl` command got ran against both `solace-fi` and `volmex-finance`.

This is a first iteration which only runs on v2 adapters for now.
It's easy to extend if we're happy with the idea.